fix: split test-summary comment into separate job with write perms
Fork PRs get a read-only GITHUB_TOKEN for the entire build job because it
checks out and runs untrusted fork code. Posting the PR comment from that
same job always 403s on fork PRs. Split into a build job (read-only,
writes test-summary.json artifact) and a comment job (checks out only the
base ref, never fork code, so it's safe to grant pull-requests: write).

Fixes #324
